requirements.txt
streamlit
pandas
psycopg2-binary
kafka-python
praw
pyspark
prometheus_client
loguru
